Use mpy map and heap where possible, use root pointer
This replaces the use of std::map by micropython heap based maps.
As a result, we need a root pointer for the root map such that the
mpy garbage collector does not collect the micropython-wrap objects.

This root pointer needs to be declared once outside of micropython-wrap
such that it is generated by the micropython generators.

All objects referenced by the map should also be on the
micropython-heap. In this way, the garbage collector can properly
track lifetimes.

This is the first step into allowing us to use micropython-wrap on mcus,
and also allows us to reset the micropython-wrap state by resetting the
global state map.